Add terminal UI components and update CLI

+++ src/cli/cli.ts
@@ -2,6 +2,10 @@ import { Command } from "commander";
 import { saveWeather, listWeather, deleteWeather } from "../utils/storage";
 import { getWeather } from "../services/weather";
 import { getLocation } from "../services/location";
+import ora from "ora";
+import { renderWeatherCard } from "../ui/weatherCard";
+import inquirer from "inquirer";
+import { handleError } from "../utils/error";
 
 export function parseCLI() {
     const program = new Command();
@@ -51,27 +55,47 @@ export function parseCLI() {
 }
 
 async function handleGet(options: any) {
-    const location = await getLocation(options.city);
-    const weatherData = await getWeather(
-        location.city,
-        location.lat,
-        location.lon,
-        options.date
-    );
+    const spinner = ora("Fetching weather data...").start();
 
-    console.log(weatherData);
+    try {
+        const location = await getLocation(options.city);
+        const weatherData = await getWeather(
+            location.city,
+            location.lat,
+            location.lon,
+            options.date
+        );
 
-    if (options.save) {
-        saveWeather(weatherData);
-        console.log("Saved successfully.");
+        renderWeatherCard({
+            city: location.city,
+            country: location.country,
+            temperature: weatherData.temperature,
+            windSpeed: weatherData.windSpeed,
+            date: weatherData.date
+        });
+
+        spinner.succeed("Weather data fetched successfully.");
+
+        if (options.save) 
+            saveWeather(weatherData);
+    } catch (err: any) {
+        spinner.fail("Failed to fetch weather data.");
+        handleError(err);
     }
 }
 
 function handleList(options: any) {
-    console.log("FILTERS:", options);
     listWeather(options);
 }
 
-function handleDelete(options: any) {
+async function handleDelete(options: any) {
+    const answer = await inquirer.prompt([
+        {
+            type: "confirm",
+            name: "confirm",
+            message: "Are you sure you want to delete these records?",
+        }
+    ]);
+    if (!answer.confirm) return;
     deleteWeather(options);
 }
